+++ CHANGELOG.md
@@ -2,6 +2,19 @@
 
 ## [Unreleased]
 
+## [2.0.0]
+
+### Added
+
+- All settings are now done through the Tool Window
+  - Quick Controls Panel: quickly toggle features through toggle buttons
+  - Settings Panel: fine-grained control over Window Accent configuration
+
+### Removed
+
+- The configuration settings in the Settings view have been moved to the Tool Window
+  - Settings now have a short paragraph informing the user about the Tool Window
+
 ## [1.6.3]
 
 ### Fixed
@@ -705,7 +718,8 @@
 - Window color management
 - Title numbering options
 
-[Unreleased]: https://github.com/alexBlakeGoudemond/jetbrains-window-accent/compare/1.6.3...HEAD
+[Unreleased]: https://github.com/alexBlakeGoudemond/jetbrains-window-accent/compare/2.0.0...HEAD
+[2.0.0]: https://github.com/alexBlakeGoudemond/jetbrains-window-accent/compare/1.6.3...2.0.0
 [1.6.3]: https://github.com/alexBlakeGoudemond/jetbrains-window-accent/compare/1.6.2...1.6.3
 [1.6.2]: https://github.com/alexBlakeGoudemond/jetbrains-window-accent/compare/1.6.1...1.6.2
 [1.6.1]: https://github.com/alexBlakeGoudemond/jetbrains-window-accent/compare/1.6.0...1.6.1